% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorglm.R
\name{anchorglm}
\alias{anchorglm}
\title{Fitting Anchored Generalized Linear Models}
\usage{
anchorglm(
  formula,
  A_formula,
  data,
  xi,
  m = 1,
  family = gaussian,
  type = c("deviance", "pearson")
)
}
\arguments{
\item{formula}{an object of class \code{"\link{formula}"} for the response
and covariate variables.}

\item{A_formula}{an object of class \code{"\link{formula}"} for the anchor
variables.}

\item{data}{an optional data frame, list or environment containing the
variables in the model. If not found in data, the variables are taken from
\code{environment(formula)}.}

\item{xi}{a numeric value for the hyperparameter xi.}

\item{m}{number of independent bernoulli trials. See the vignette with the
binomial example for help.}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
See \code{\link{family}} for details. The families supported up to now are
\code{gaussian}, \code{binomial} and \code{poisson}.}

\item{type}{the type of residuals that should be used for the anchor
penalty. The supported alternatives are \code{"deviance"} (default) and
\code{"pearson"}. Can be abbreviated.}
}
\value{
\code{anchorglm} returns an object of class \code{"anchorglm"}.\cr\cr
The function \code{\link{summary}} (i.e., \code{\link{summary.anchorglm}})
can be used to obtain or print a summary of the results.\cr\cr
The generic accessor functions \code{logLik}, \code{coef}, \code{predict} and \code{residuals}
can be used to extract various useful features of the value returned by
\code{anchorglm}.\cr\cr
An object of class \code{"anchorglm"} is a list containing at least the
following components:
\item{formula}{the response and covariate \code{"\link{formula}"} used.}
\item{A_formula}{the anchor \code{"\link{formula}"} used.}
\item{data}{the \code{data} argument.}
\item{xi}{the hyperparameter used.}
\item{m}{the used bernoulli trials coefficients. If not used set to \code{1}.}
\item{family}{the \code{"\link{formula}"} object used.}
\item{logLik}{the log-likelihood function used corresponding to the used
family.}
\item{devianceRes}{the deviance residuals function used corresponding to
the used family.}
\item{pearsonRes}{the pearson residuals function used corresponding to the
used family.}
\item{optim}{the \code{\link{optim}} output as a list.}
\item{coefficients}{the optimal coefficients derived by optim.}
}
\description{
\code{anchorglm} is used to fit anchored generalized linear models, specified by
giving an anchor variable, a symbolic description of the linear predictor
and a description of the error distribution.
}
\details{
The response-covariate formula must have the form
\code{response ~ covariates}, where \code{response} is the response vector and
\code{covariates} is a series of terms which specifies a linear predictor for
the response.\cr\cr
For binomial families the response can also be specified as a two-column
matrix with the columns giving the numbers of successes and failures.\cr\cr
The anchor formula must have the form \code{~ anchors}, where \code{anchors} contains
all given anchor variables.
}
