% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-likelihoods_and_residuals.R
\name{binary_deviance}
\alias{binary_deviance}
\title{Binary deviance residuals}
\usage{
binary_deviance(b, Y, X, linkinv, m, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{linkinv}{inversion of link function}

\item{m}{integer}

\item{...}{further arguments passed to or from other methods.}
}
\value{
deviance residuals
}
\description{
Binary deviance residuals
}
