% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-residuals-penalties.R
\name{binary_deviance_res}
\alias{binary_deviance_res}
\title{Binary deviance residuals}
\usage{
binary_deviance_res(b, Y, X, linkinv, m, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{linkinv}{inversion of link function}

\item{m}{integer}

\item{...}{other}

\item{A}{nXq matrix}
}
\value{
deviance residuals
}
\description{
Binary deviance residuals
}
