% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-and-penalties.R
\name{binary_likelihood}
\alias{binary_likelihood}
\title{Binary log likelihood}
\usage{
binary_likelihood(b, Y, X, m)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{m}{integer}
}
\value{
log likelihood objective
}
\description{
Binary log likelihood
}
