% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-residuals-penalties.R
\name{binary_penalty}
\alias{binary_penalty}
\title{Binary anchor penalty}
\usage{
binary_penalty(b, Y, X, A, linkinv, m, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{A}{nXq matrix}

\item{linkinv}{inversion of link function}

\item{m}{integer}

\item{...}{other}
}
\value{
anchor penalty objective
}
\description{
Binary anchor penalty
}
