% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.anchorglm}
\alias{coef.anchorglm}
\title{coefficients of anchor glm object}
\usage{
\method{coef}{anchorglm}(object, ...)
}
\arguments{
\item{object}{anchor glm object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric
}
\description{
Returns coefficientes of a anchor glm object
}
