% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_dwt.R
\name{Energy}
\alias{Energy}
\title{Compute relative energy for a general wavelet coefficient matrix.}
\usage{
Energy(D)
}
\arguments{
\item{D}{An N x T matrix of wavelet coefficients.}
}
\value{
En A N x T matrix containing the relative energy, such that row i and
column k entry contains the cumulative proportion of energy explained by this
coefficient and all coefficients greater in absolute value than it for subject i.
}
\description{
Compute relative energy for a general wavelet coefficient matrix.
}
