% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flf_basissel_ae.R
\name{flf_basissel_ae}
\alias{flf_basissel_ae}
\title{Basis selection for AE.}
\usage{
flf_basissel_ae(
  mat,
  kf,
  lim = lim,
  incr = incr,
  ae_args = list(),
  verbose = TRUE
)
}
\arguments{
\item{Y}{an n times p data matrix.}
}
\description{
Basis selection for AE.
}
