% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flf_basissel_dwt.2d.R
\name{flf_basissel_dwt.2d}
\alias{flf_basissel_dwt.2d}
\title{Assess losslessness of 2-D DWT latent feature representation.}
\usage{
flf_basissel_dwt.2d(
  mat,
  kf,
  lim = min(ncol(mat) - 1, nrow(mat) - 1),
  incr = 1,
  verbose = TRUE
)
}
\arguments{
\item{mat}{an n times p times q data array, containing p times q images in the n slices}

\item{kf}{k, the number of folds in k-fold cross-validation,}

\item{center}{Whether or not the data should be mean-centered prior to performing the latent feature transformation.}

\item{scale}{Whether the data should be scaled prior to analysis, so that each variable has standard deviation 1.}
}
\description{
Assess losslessness of 2-D DWT latent feature representation.
}
