% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flf_basissel_pca.R
\name{flf_basissel_pca}
\alias{flf_basissel_pca}
\title{Assess losslessness of PCA latent feature representation.}
\usage{
flf_basissel_pca(
  mat,
  kf,
  lim = min(ncol(mat) - 1, nrow(mat) - 1),
  incr = 1,
  center = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{mat}{an n times p data matrix.}

\item{kf}{k, the number of folds in k-fold cross-validation,}

\item{center}{Whether or not the data should be mean-centered prior to performing the latent feature transformation.}

\item{scale}{Whether the data should be scaled prior to analysis, so that each variable has standard deviation 1.}
}
\description{
Assess losslessness of PCA latent feature representation.
}
