% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_example_data.R
\name{generate_example_data}
\alias{generate_example_data}
\title{Generate example data for glare}
\usage{
generate_example_data(
  n,
  dim_X = 1,
  dim_A = 1,
  dim_H = 1,
  family = gaussian,
  m_max = 1,
  A_distr = c("normal", "rademacher"),
  XH = NULL,
  XA = NULL,
  HX = NULL,
  HA = NULL,
  GX = NULL,
  GH = NULL,
  GA = NULL,
  ...
)
}
\arguments{
\item{n}{number of observations.}

\item{dim_X}{number of covariates.}

\item{dim_A}{number of anchor variables.}

\item{dim_H}{number of hidden confounders.}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
See \code{\link{family}} for details. The families supported up to now are
\code{gaussian}, \code{binomial} and \code{poisson}.}

\item{m_max}{maximal number of indepdent bernoulli trials, if binomial family
is used.}

\item{A_distr}{distribution from which the Anchor variable is generated.}

\item{XH}{coefficients of how X depends on H as a dim_X times dim_H matrix.}

\item{XA}{coefficients of how X depends on A as a dim_X times dim_A matrix.}

\item{HX}{coefficients of how H depends on X as a dim_H times dim_X matrix.}

\item{HA}{coefficients of how H depends on A as a dim_H times dim_A matrix.}

\item{GX}{coefficients of how G depends on X as a dim_X vector.}

\item{GH}{coefficients of how G depends on H as a dim_H vector.}

\item{GA}{coefficients of how G depends on A as a dim_A vector.}

\item{...}{other}
}
\value{
returns a data list of matrices of all the constructed variables.
}
\description{
Generate example data for glare
}
\details{
For this function only one direction of dependence between X and H
are allowed.\cr\cr
The response Y is generated with the corresponding family, linkfunction and
the variable G, e.g.\cr \code{Y = rpois(n = n, lambda = linkinv(G))}.\cr\cr
As input the desired dimension of the variables are used for a weak
check for correct input of the variable coefficients.\cr\cr
For a script to generate data manipulate the given example below.
}
\examples{
 # H = epsH
 HX = NULL
 HA = NULL

 # X1 = H + 0.5 * A1 - 0.2 * A2 + epsX
 # X2 = H + 0.5 * A1 - 0.4 * A2 + epsX
 XH = rbind(1, 1)
 XA = rbind(c(0.5, -0.2), c(0.5, -0.4))

 # G = 3 * X1 + 3 * X2 + H - 2 * A1
 GX = c(3, 3)
 GH = 1
 GA = c(-2, 0)

 generate_example_data(n = 1000, dim_X = 2, dim_A = 2, dim_H = 1,
                       family = binomial, m_max = 5,
                       A_distr = "rademacher",
                       XH = XH, XA = XA,
                       HX = HX,  HA = HA,
                       GX = GX, GH = GH, GA = GA)
}
