% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt-work.R
\name{get_Energy_scree}
\alias{get_Energy_scree}
\title{Compute scree from relative energy matrix.}
\usage{
get_Energy_scree(D)
}
\arguments{
\item{D}{An N x T matrix of wavelet coefficients.}
}
\value{
En A N x T matrix containing the relative energy, such that row i and
column k entry contains the cumulative proportion of energy explained by this
coefficient and all coefficients greater in absolute value than it.
}
\description{
Compute scree from relative energy matrix.
}
