% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{get_one_minus_squared_correlation}
\alias{get_one_minus_squared_correlation}
\title{Compute 1 - Squared Correlation}
\usage{
get_one_minus_squared_correlation(observed, predicted)
}
\arguments{
\item{observed}{Numeric vector of observed values.}

\item{predicted}{Numeric vector of predicted values.}
}
\value{
Numeric value representing 1 - squared correlation.
}
\description{
Calculates 1 minus the squared correlation between observed and predicted values.
If `predicted` values are constant, the squared correlation is set to 0.
}
\examples{
get_one_minus_squared_correlation(c(1, 2, 3), c(1, 2, 2))
}
