% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlaRe-package.R
\docType{package}
\name{GLaRe-package}
\alias{GLaRe-package}
\title{GLaRe: Graphical Assessment of Latent Feature Representations}
\description{
Implements methods for assessing the losslessness of latent feature representations for complex object-valued data using out-of-sample performance. The package provides tools for computing losslessness measures and for visualizing results through graphical displays, including an interactive Shiny-based interface.
}
\author{
\strong{Maintainer}: Edward Gunning \email{edward.gunning@pennmedicine.upenn.edu}

Authors:
\itemize{
  \item Emma Zohner \email{etraore@gmail.com}
}

}
\keyword{internal}
