% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glaucoma_data}
\alias{glaucoma_data}
\title{Glaucoma data}
\format{
A data frame with 306 rows and 14400 variables.
}
\usage{
glaucoma_data
}
\description{
A dataset used to illustrate latent feature reconstruction methods.
}
\keyword{datasets}
