% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_dwt.2d.R
\name{learn_dwt.2d}
\alias{learn_dwt.2d}
\title{Learning function for 2-D DWT.}
\usage{
learn_dwt.2d(Y, p1, p2)
}
\arguments{
\item{Y}{an n times p data matrix.}
}
\description{
Learning function for 2-D DWT.
}
