% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flf_basissel_dwt.R
\name{learn_dwt}
\alias{learn_dwt}
\title{learn}
\usage{
learn_dwt(
  Y,
  wavespecs = list("d4", n.levels = 5, boundary = "periodic"),
  breaks = c(seq(0.8, 0.875, by = 0.025), seq(0.9, 0.95, by = 0.01), seq(0.95, 1, by =
    4e-04))
)
}
\description{
learn
}
