% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flf_basissel_pca.R
\name{learn_pca}
\alias{learn_pca}
\title{Helper function to perform principal component analysis.}
\usage{
learn_pca(Y)
}
\arguments{
\item{Y}{an n times p data matrix.}
}
\description{
Helper function to perform principal component analysis.
}
