% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.anchorglm}
\alias{logLik.anchorglm}
\title{log likelihood of a anchor glm object}
\usage{
\method{logLik}{anchorglm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{anchor glm object}

\item{newdata}{used for test data, default is NULL}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric
}
\description{
Returns log likelihood of a anchor glm object using train or new data
}
