% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.glare}
\alias{logLik.glare}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{glare}(object, newdata = NULL, parameter = NULL, indiv = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{"\link{glare}"}.}

\item{newdata}{an optional data frame, list or environment containing the
variables in the model. If not used, the variables are taken from the
objective.}

\item{parameter}{optional parameter input. Default takes the parameter
from the glare objective.}

\item{indiv}{logical input if log-likelihood of each observation or the sum
of the log-likelihoods should be returned.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns the log-likelihood of an object of class \code{"glare"}.
}
\description{
Returns log-likelihood of object with class \code{"glare"}, using initial
training or new data.
}
