% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-likelihoods_and_residuals.R
\name{normal_deviance}
\alias{normal_deviance}
\title{Normal deviance residuals}
\usage{
normal_deviance(b, Y, X, linkinv, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{linkinv}{inversion of link function}

\item{...}{further arguments passed to or from other methods.}
}
\value{
deviance residuals
}
\description{
Normal deviance residuals
}
