% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-and-penalties.R
\name{normal_penalty}
\alias{normal_penalty}
\title{Normal anchor penalty}
\usage{
normal_penalty(b, Y, X, m, P.A)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{m}{integer}

\item{P.A}{orthogonal projection matrix of A}
}
\value{
anchor penalty objective
}
\description{
Normal anchor penalty
}
