% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_1D_reconstruction}
\alias{plot_1D_reconstruction}
\title{Plot 1D Reconstruction}
\usage{
plot_1D_reconstruction(GLaRe_output, Y)
}
\arguments{
\item{GLaRe_output}{The object returned from a call to the `GLaRe()` function.}

\item{Y}{An n x p matrix containing n observations to reconstruct.}
}
\value{
A plot comparing the original and reconstructed 1D data.
}
\description{
Visualizes original and reconstructed 1D data (e.g., time series).
}
