% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_eye_reconstruction}
\alias{plot_eye_reconstruction}
\title{Plot a reconstruction of the eye data.}
\usage{
plot_eye_reconstruction(GLaRe_output, y)

plot_eye_reconstruction(GLaRe_output, y)
}
\description{
Plot a reconstruction of the eye data.

Plot a reconstruction of the glaucoma data.
}
