% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_gel}
\alias{plot_gel}
\title{Plot Gel Data}
\usage{
plot_gel(y)
}
\arguments{
\item{y}{A 646 x 861 matrix representing gel data.}
}
\value{
A heatmap of the gel data.
}
\description{
Visualizes proteomic gels data.
}
