% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_gel_reconstruction}
\alias{plot_gel_reconstruction}
\title{Plot a reconstruction of gels data.}
\usage{
plot_gel_reconstruction(GLaRe_output, y)
}
\arguments{
\item{GLaRe_output}{the object returned from a call to the `GLaRE()` function.}

\item{Y}{An n times p matrix containing the n observations for which you want their reconstruction displayed.}
}
\description{
Plot a reconstruction of gels data.
}
