% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_mnist}
\alias{plot_mnist}
\title{Plot a Single MNIST Image}
\usage{
plot_mnist(y, main = NULL)
}
\arguments{
\item{y}{A 28x28 numeric matrix representing an MNIST image.}

\item{main}{Title for the plot. Defaults to NULL.}
}
\value{
A visual plot of the MNIST image.
}
\description{
Displays an individual MNIST image.
}
