% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_mnist_reconstruction}
\alias{plot_mnist_reconstruction}
\title{Plot a reconstruction of the mnist data.}
\usage{
plot_mnist_reconstruction(GLaRe_output, y)
}
\description{
Plot a reconstruction of the mnist data.
}
