% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-likelihoods_and_residuals.R
\name{poisson_deviance}
\alias{poisson_deviance}
\title{Poisson deviance residuals}
\usage{
poisson_deviance(b, Y, X, linkinv, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{linkinv}{inversion of link function}

\item{...}{other}
}
\value{
deviance residuals
}
\description{
Poisson deviance residuals
}
