% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-residuals-penalties.R
\name{poisson_deviance_res}
\alias{poisson_deviance_res}
\title{Poisson deviance residuals}
\usage{
poisson_deviance_res(b, Y, X, A, linkinv, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{A}{nXq matrix}

\item{linkinv}{inversion of link function}

\item{...}{other}
}
\value{
deviance residuals
}
\description{
Poisson deviance residuals
}
