% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-likelihoods_and_residuals.R
\name{poisson_likelihood}
\alias{poisson_likelihood}
\title{Poisson likelihood}
\usage{
poisson_likelihood(b, Y, X, linkinv, indiv = FALSE, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{linkinv}{inversion of link function}

\item{indiv}{logical input if log-likelihood of each observation or the sum
of the log-likelihoods should be returned.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
log likelihood objective
}
\description{
Poisson likelihood
}
