% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-residuals-penalties.R
\name{poisson_penalty}
\alias{poisson_penalty}
\title{Poisson anchor penalty}
\usage{
poisson_penalty(b, Y, X, A, linkinv, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{A}{nXq matrix}

\item{linkinv}{inversion of link function}

\item{...}{other}
}
\value{
anchor penalty objective
}
\description{
Poisson anchor penalty
}
