% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.glare}
\alias{predict.glare}
\title{Model Predictions}
\usage{
\method{predict}{glare}(
  object,
  newdata = NULL,
  parameter = NULL,
  type = c("link", "response"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{"\link{glare}"}.}

\item{newdata}{an optional data frame, list or environment containing the
variables in the model. If not used, the variables are taken from the
objective.}

\item{parameter}{optional parameter input. Default takes the parameter
from the glare objective.}

\item{type}{the scale of prediction. The default is on the scale of
the linear predictors; the alternative \code{"response"} is on the scale of the
response variable. Thus for a default binomial model the default predictions
are of log-odds (probabilities on logit scale) and \code{type = "response"} gives
the predicted probabilities.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The returned predictions depend on the choosen type of scale (see
above).
}
\description{
Returns predictions on scale of linear predictor or response of \code{"glare"}
object using train or new data.
}
