% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_dwt.R
\name{prepare_pad_dwt}
\alias{prepare_pad_dwt}
\title{Pad a signal to dyadic length for 1-D DWT with 0's each side.}
\usage{
prepare_pad_dwt(Y)
}
\arguments{
\item{Y}{an N times T data matrix containing N replicate observations of a T-dimensional signal in its rows.}
}
\description{
Pad a signal to dyadic length for 1-D DWT with 0's each side.
}
