% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.glare}
\alias{print.glare}
\title{Object Summary}
\usage{
\method{print}{glare}(x, show_hess = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"\link{glare}"}.}

\item{show_hess}{should the approximated hessian be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints an object of class \code{"glare"} in compressed form.
}
\description{
Returns summary of an object of class \code{"glare"}.
}
