% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.glare}
\alias{print.summary.glare}
\title{Print of summary method for \code{"glare"} object.}
\usage{
\method{print}{summary.glare}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{"\link{glare}"}.}

\item{digits}{integer for number of digits used for the rounded result.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints summary of a \code{"glare"} object.
}
\description{
Returns summary of a glare object.
}
