% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{rep.col}
\alias{rep.col}
\title{Repeat a Column Vector}
\usage{
\method{rep}{col}(x, n)
}
\arguments{
\item{x}{Numeric vector to repeat.}

\item{n}{Integer specifying the number of repetitions.}
}
\value{
A matrix where `x` is repeated `n` times as columns.
}
\description{
Creates a matrix by repeating a column vector `x` `n` times.
}
\examples{
rep.col(c(1, 2, 3), 3)
}
