% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{rep_col}
\alias{rep_col}
\title{Repeat a Column Vector}
\usage{
rep_col(x, n)
}
\arguments{
\item{x}{Numeric vector to repeat.}

\item{n}{Integer specifying the number of repetitions.}
}
\value{
A matrix where `x` is repeated `n` times as columns.
}
\description{
Creates a matrix by repeating a column vector `x` `n` times.
}
