% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.anchorglm}
\alias{residuals.anchorglm}
\title{residuals of a anchor glm object}
\usage{
\method{residuals}{anchorglm}(object, newdata = NULL, type = c("deviance", "pearson"), ...)
}
\arguments{
\item{object}{anchor glm object}

\item{newdata}{used for test data, default is NULL}

\item{type}{the type of residuals which should be returned. The alternatives are: "deviance" (default) and "pearson". Can be abbreviated.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric
}
\description{
Returns deviance or pearson residuals of a anchor glm object using train or new data
}
