% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.glare}
\alias{residuals.glare}
\title{Extract Model Residuals}
\usage{
\method{residuals}{glare}(
  object,
  newdata = NULL,
  parameter = NULL,
  type = c("deviance", "pearson", "classical"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{"\link{glare}"}.}

\item{newdata}{an optional data frame, list or environment containing the
variables in the model. If not used, the variables are taken from the
objective.}

\item{parameter}{optional parameter input. Default takes the parameter
from the glare objective.}

\item{type}{the type of residuals which should be returned. The alternatives
are: \code{"deviance"} (default), \code{"pearson"} and \code{"classical"}. Can be abbreviated.
Please use \code{"classical"} only for a gaussian setup.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Residuals extracted from the model object of class \code{"glare"}.
}
\description{
Returns deviance or pearson residuals of a \code{"glare"} object using train or
new data.
}
