% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.glare}
\alias{summary.glare}
\title{Object Summary}
\usage{
\method{summary}{glare}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{"\link{glare}"}.}

\item{digits}{integer for number of digits used for the rounded result.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns object of class \code{"summary.glare"}.
}
\description{
Returns summary of an object of class \code{"glare"}.
}
