% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_dwt.R
\name{waveslim_dwt}
\alias{waveslim_dwt}
\title{Apply DWT to a 1-D signal and return the wavelet coefficients as a list.}
\usage{
waveslim_dwt(y)
}
\arguments{
\item{y}{a vector conatining the 1-D signal, must be of dyadic length.}
}
\description{
Apply DWT to a 1-D signal and return the wavelet coefficients as a list.
}
