% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_dwt.R
\name{waveslim_dwt_to_mat}
\alias{waveslim_dwt_to_mat}
\title{Apply DWT to a a matrix of n 1-D signals and return the wavelet coefficients as a matrix.}
\usage{
waveslim_dwt_to_mat(Y)
}
\arguments{
\item{Y}{an N times T data matrix containing N replicate observations of a T-dimensional signal in its rows. T must be dyadic length.}
}
\description{
Apply DWT to a a matrix of n 1-D signals and return the wavelet coefficients as a matrix.
}
