% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_glm.R
\name{anchor_glm}
\alias{anchor_glm}
\title{Main script for AGLM}
\usage{
anchor_glm(
  formula,
  A.formula,
  xi,
  m = 1,
  family = gaussian,
  type = c("deviance", "pearson")
)
}
\arguments{
\item{formula}{an object of class "formula" for the response and covariate variables.}

\item{A.formula}{an object of class "formula" for the anchor variables.}

\item{xi}{numeric}

\item{m}{vector}

\item{family}{character}

\item{type}{the type of residuals that should be used for the anchor penalty. The alternatives are: "deviance" (default) and "pearson". Can be abbreviated.}
}
\value{
numeric
}
\description{
Script that constructs Anchor GLM objective and optimizes it
using optim from stats.
}
