% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_glm.R
\name{anchor_glm}
\alias{anchor_glm}
\title{Anchor Generalized Linear Model main script}
\usage{
anchor_glm(Y, X, A, xi, m, family)
}
\arguments{
\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{A}{nxq matrix}

\item{xi}{numeric}

\item{m}{integer}

\item{family}{character}
}
\value{

}
\description{
Anchor Generalized Linear Model main script
}
\examples{
#example to be submitted
}
