% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor_glm.R
\name{anchor_glm}
\alias{anchor_glm}
\title{Main script for AGLM}
\usage{
anchor_glm(Y, X, A, xi, m = 0, family = gaussian)
}
\arguments{
\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{A}{nxq matrix}

\item{xi}{numeric}

\item{m}{integer}

\item{family}{character}
}
\value{
numeric
}
\description{
SCript that constructs Anchor GLM objective and optimizes it
using optim from stats.
}
\examples{
Y <- c(0,1,0)
X <- matrix(c(1,2,3, 11,12,13), nrow = 3, ncol = 2)
A <- matrix(c(1,-1,-1, 1,-1,1), nrow = 3, ncol = 2)
anchor_glm(Y, X, A, 2, 1, "binomial")
}
