% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.glare}
\alias{coef.glare}
\title{Extract Model Coefficients}
\usage{
\method{coef}{glare}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"\link{glare}"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Coefficients extracted from the model object of class \code{"glare"}.
}
\description{
Returns model coefficients of object with class \code{"glare"}.
}
