% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{distribution_plot}
\alias{distribution_plot}
\title{Plot the full distribution of individual validation losses.}
\usage{
distribution_plot(GLaRe_output, jitter = 0.2, cex = 1)
}
\arguments{
\item{GLaRe_output}{the object returned from a call to the `GLaRE()` function.}
}
\description{
Plot the full distribution of individual validation losses.
}
