% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flf_basissel.R
\name{flf_basissel}
\alias{flf_basissel}
\title{Assess losslessness of general feature representation.}
\usage{
flf_basissel(
  mat,
  learn,
  ae_args,
  kf,
  latent_dim_from = 1,
  latent_dim_to = min(ncol(mat), nrow(mat) - 1),
  latent_dim_by = 1,
  loss_function = get_one_minus_squared_correlation,
  learn_function = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mat}{an n times p data matrix.}

\item{kf}{k, the number of folds in k-fold cross-validation,}

\item{center}{Whether or not the data should be mean-centered prior to performing the latent feature transformation.}

\item{scale}{Whether the data should be scaled prior to analysis, so that each variable has standard deviation 1.}
}
\description{
Assess losslessness of general feature representation.
}
