% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flf_basissel_dwt.R
\name{flf_basissel_dwt}
\alias{flf_basissel_dwt}
\title{DWT}
\usage{
flf_basissel_dwt(
  mat,
  kf,
  learn = learn_dwt,
  wavespecs = list(filter = "d4", n.levels = 5, boundary = "periodic"),
  center = TRUE,
  breaks = c(seq(0.8, 0.975, by = 0.025), seq(0.975, 1, by = 0.001))
)
}
\description{
DWT
}
