% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glare.R
\name{glare}
\alias{glare}
\title{Fitting Generalized Linear Anchor Regression}
\usage{
glare(
  formula,
  A_formula,
  data,
  xi,
  family = gaussian,
  type = c("deviance", "pearson", "classical")
)
}
\arguments{
\item{formula}{an object of class \code{"\link{formula}"} for the response
and covariate variables.}

\item{A_formula}{an object of class \code{"\link{formula}"} for the anchor
variables.}

\item{data}{an optional data frame, list or environment containing the
variables in the model. If not found in data, the variables are taken from
\code{environment(formula)}.}

\item{xi}{a numeric value for the hyperparameter xi.}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a family
function, a family function or the result of a call to a family function.
See \code{\link{family}} for details. The families supported up to now are
\code{gaussian}, \code{binomial} and \code{poisson}.}

\item{type}{the type of residuals that should be used for the anchor
penalty. The supported alternatives are \code{"deviance"} (default), \code{"pearson"}
and \code{"classical"}. Can be abbreviated. Please use \code{"classical"} only for a
gaussian setup.}
}
\value{
\code{glare} returns an object of class \code{"glare"}.\cr\cr
The function \code{\link{summary}} (i.e., \code{\link{summary.glare}})
can be used to obtain or print a summary of the results.\cr\cr
The generic accessor functions \code{logLik}, \code{coef}, \code{predict} and \code{residuals}
can be used to extract various useful features of the value returned by
\code{glare}.\cr\cr
An object of class \code{"glare"} is a list containing at least the
following components:
\item{formula}{the response and covariate \code{"\link{formula}"} used.}
\item{A_formula}{the anchor \code{"\link{formula}"} used.}
\item{data}{the \code{data} argument as data.frame.}
\item{xi}{the hyperparameter used.}
\item{family}{the \code{"\link{formula}"} object used.}
\item{logLik}{the log-likelihood function used corresponding to the used
family.}
\item{devianceRes}{the deviance residuals function used corresponding to
the used family.}
\item{pearsonRes}{the pearson residuals function used corresponding to the
used family.}
\item{optim}{the \code{\link{optim}} output as a list.}
\item{coefficients}{the optimal coefficients derived by optim.}
}
\description{
\code{glare} is used to fit Generalized Linear Anchor REgression (GLARE), specified by
giving an anchor variable, a symbolic description of the linear predictor
and a description of the error distribution.
}
\details{
The response-covariate formula must have the form
\code{response ~ covariates}, where \code{response} is the response vector and
\code{covariates} is a series of terms which specifies a linear predictor for
the response.\cr\cr
For binomial families the response can also be specified as a two-column
matrix with the columns giving the numbers of successes and failures. If the
number of independent bernoulli trials is provided trough the data
argument, the variable has be called \code{m}. See the vignette with the
binomial example for help.\cr\cr
The anchor formula must have the form \code{~ anchors}, where \code{anchors} contains
all given anchor variables.\cr\cr
For a classical linear gaussian setup we recommand to use anchor regression,
constructed by Rothenhaeusler et al. (see \code{?anchor_regression} for more
information).
}
