% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_ae.R
\name{learn_ae}
\alias{learn_ae}
\title{Learning function for autoencoder (AE)}
\usage{
learn_ae(Y, k, ae_args)
}
\arguments{
\item{Y}{An n times p data matrix.}

\item{k}{The latent feature dimension. Also known as the "bottleneck dimension" in machine learning terminology.}

\item{ae_args}{A list containing the following named elements to define the architecture and training of the AE: `layer_1_dim`, `layer_2_dim`, `link_fun`, `epochs`, `loss` and `batch_size`.}
}
\description{
Learning function for autoencoder (AE)
}
