% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt.2d-work.R
\name{learn_dwt.2d}
\alias{learn_dwt.2d}
\title{2-D DWT latent feature learning function.}
\usage{
learn_dwt.2d(Y)
}
\arguments{
\item{Y}{an n times p data matrix.}
}
\description{
2-D DWT latent feature learning function.
}
