% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_pca.R
\name{learn_pca}
\alias{learn_pca}
\title{Learning function for principal component analysis (PCA).}
\usage{
learn_pca(Y)
}
\arguments{
\item{Y}{an n times p data matrix.}
}
\description{
Learning function for principal component analysis (PCA).
}
