% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.glare}
\alias{logLik.glare}
\title{log likelihood of a anchor glm object}
\usage{
\method{logLik}{glare}(object, newdata = NULL, parameter = NULL, ...)
}
\arguments{
\item{object}{anchor glm object}

\item{newdata}{used for test data, default is NULL}

\item{parameter}{parameter that should be used, default takes the parameter
from the glare objective.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric
}
\description{
Returns log likelihood of a anchor glm object using train or new data
}
