% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_1D_reconstruction}
\alias{plot_1D_reconstruction}
\title{Plot a reconstruction of 1-D (e.g., time series) data.}
\usage{
plot_1D_reconstruction(GLaRe_output, Y)
}
\arguments{
\item{GLaRe_output}{the object returned from a call to the `GLaRE()` function.}

\item{Y}{An n times p matrix containing the n observations for which you want their reconstruction displayed.}
}
\description{
Plot a reconstruction of 1-D (e.g., time series) data.
}
