% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_eye}
\alias{plot_eye}
\title{Plot Glaucoma Eye Data}
\usage{
plot_eye(y)
}
\arguments{
\item{y}{A 14400-dimensional numeric vector of measurements.}
}
\value{
A ggplot2 object visualizing the eye data.
}
\description{
Visualizes glaucoma eye data as a 2D radial projection.
}
