% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_eye_reconstruction}
\alias{plot_eye_reconstruction}
\title{Plot Eye Reconstruction}
\usage{
plot_eye_reconstruction(GLaRe_output, y)
}
\arguments{
\item{GLaRe_output}{The object returned from a call to the `GLaRe()` function.}

\item{y}{A 14400-dimensional numeric vector of measurements.}
}
\value{
A ggplot2 object displaying the original and reconstructed data.
}
\description{
Visualizes the original and reconstructed glaucoma eye data side by side.
}
