% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_gel_reconstruction}
\alias{plot_gel_reconstruction}
\title{Plot Gel Reconstruction}
\usage{
plot_gel_reconstruction(GLaRe_output, y)
}
\arguments{
\item{GLaRe_output}{The object returned from a call to the `GLaRe()` function.}

\item{y}{A 646 x 861 matrix representing gel data.}
}
\value{
A visual comparison of the original and reconstructed gel data.
}
\description{
Visualizes the original and reconstructed proteomic gels data side by side.
}
