% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_mnist}
\alias{plot_mnist}
\title{Helper function to plot a single image from the mnist data.}
\usage{
plot_mnist(y, main = NULL)
}
\description{
Helper function to plot a single image from the mnist data.
}
