% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_mnist_reconstruction}
\alias{plot_mnist_reconstruction}
\title{Plot a reconstruction of the mnist data.}
\usage{
plot_mnist_reconstruction(GLaRe_output, y)
}
\arguments{
\item{GLaRe_output}{the object returned from a call to the `GLaRE()` function.}

\item{y}{a 28 times 28 matrix representing an image from the MNIST dataset.}
}
\description{
Plot a reconstruction of the mnist data.
}
