% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_mnist_reconstruction}
\alias{plot_mnist_reconstruction}
\title{Plot MNIST Reconstruction}
\usage{
plot_mnist_reconstruction(GLaRe_output, y)
}
\arguments{
\item{GLaRe_output}{The object returned from a call to the `GLaRe()` function.}

\item{y}{A 28x28 matrix representing an MNIST image.}
}
\value{
A visual comparison of the original and reconstructed MNIST image.
}
\description{
Displays the original and reconstructed MNIST image side by side.
}
