% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{plot_train_validation_ratio}
\alias{plot_train_validation_ratio}
\title{Plot the ratio of the average training loss to the average validation loss.}
\usage{
plot_train_validation_ratio(GLaRe_output)
}
\description{
Plot the ratio of the average training loss to the average validation loss.
}
