% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-residuals-penalties.R
\name{poisson_likelihood}
\alias{poisson_likelihood}
\title{Poisson likelihood}
\usage{
poisson_likelihood(b, Y, X, ...)
}
\arguments{
\item{b}{parameter vector}

\item{Y}{n-dimensional vector}

\item{X}{nxp matrix}

\item{...}{other}
}
\value{
log likelihood objective
}
\description{
Poisson likelihood
}
