% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.glare}
\alias{predict.glare}
\title{predictions for anchor glm object}
\usage{
\method{predict}{glare}(
  object,
  newdata = NULL,
  parameter = NULL,
  type = c("link", "response"),
  ...
)
}
\arguments{
\item{object}{anchor glm object}

\item{newdata}{used for test data, default is NULL}

\item{parameter}{parameter that should be used, default takes the parameter
from the glare objective.}

\item{type}{the type of prediction required. The default is on the scale of
the linear predictors; the alternative "response" is on the scale of the
response variable. Thus for a default binomial model the default predictions
are of log-odds (probabilities on logit scale) and type = "response" gives
the predicted probabilities.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric
}
\description{
Returns prediction on scale of linear predictor or response of a anchor glm
object using train or new data
}
