% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.glare}
\alias{print.summary.glare}
\title{Print of summary method for \code{"glare"} object.}
\usage{
\method{print}{summary.glare}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"\link{glare}"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints summary of a \code{"glare"} object.
}
\description{
Returns summary of a glare object.
}
