% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.glare}
\alias{residuals.glare}
\title{residuals of a anchor glm object}
\usage{
\method{residuals}{glare}(
  object,
  newdata = NULL,
  parameter = NULL,
  type = c("deviance", "pearson"),
  ...
)
}
\arguments{
\item{object}{anchor glm object}

\item{newdata}{used for test data, default is NULL}

\item{parameter}{parameter that should be used, default takes the parameter
from the glare objective.}

\item{type}{the type of residuals which should be returned. The alternatives
are: "deviance" (default) and "pearson". Can be abbreviated.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
numeric
}
\description{
Returns deviance or pearson residuals of a anchor glm object using train or
new data
}
